% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{angle2hjust}
\alias{angle2hjust}
\title{Convert angles to ggplot2 hjust}
\usage{
angle2hjust(theta, multiplier = 1.5, as_degrees = FALSE)
}
\arguments{
\item{theta}{angle in radians}

\item{multiplier}{distance from point}

\item{as_degrees}{use degrees instead of radians}
}
\value{
numeric
}
\description{
Convert angles to ggplot2 hjust
}
