% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{attach_function}
\alias{attach_function}
\title{Converts the default values of a function's arguments to variables and attaches them to the global environment}
\usage{
attach_function(f)
}
\arguments{
\item{f}{Function}
}
\value{
Attaches function arguments to global environment
}
\description{
Converts the default values of a function's arguments to variables and attaches them to the global environment
}
\examples{
my_function <- function(x, y = 2) x + y

# Sets y to 2
attach_function(my_function)
}
