% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{composite_covariance}
\alias{composite_covariance}
\title{Computes covariances of composite scores given a covariance matrix and a weight matrix}
\usage{
composite_covariance(Sigma, w, correlation = FALSE)
}
\arguments{
\item{Sigma}{Covariance matrix}

\item{w}{Weight matrix. Must have the same number of rows as R}

\item{correlation}{If TRUE, return correlations instead of covariances}
}
\description{
Computes covariances of composite scores given a covariance matrix and a weight matrix
}
\examples{
Sigma <- matrix(0.6, nrow = 4, ncol = 4)
diag(Sigma) <- 1
w <- matrix(c(1,0,
              1,0,
              0,1,
              0,1),
              nrow = 4, ncol = 2)
composite_covariance(Sigma, w)
}
