% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cor2pcor}
\alias{cor2pcor}
\title{Converts a correlation matrix to a partial correlation matrix}
\usage{
cor2pcor(R)
}
\arguments{
\item{R}{correlation matrix}
}
\value{
Partial correlation matrix
}
\description{
Converts a correlation matrix to a partial correlation matrix
}
\examples{
R <- matrix(.6, nrow = 3, ncol = 3)
diag(R) <- 1
cor2pcor(R)
}
