% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cor_ellipse}
\alias{cor_ellipse}
\title{Generate correlation ellipse data}
\usage{
cor_ellipse(
  r = 0,
  mean = c(0, 0),
  sd = c(1, 1),
  p = 0.95,
  split_x = NULL,
  split_y = NULL,
  n_points = 10000
)
}
\arguments{
\item{r}{Correlation coefficient}

\item{mean}{Vector of 2 means}

\item{sd}{Vector of 2 standard deviations}

\item{p}{Proportion of data ellipse covers}

\item{split_x}{Split ellipse at x value}

\item{split_y}{Split ellipse at y value}

\item{n_points}{Number of points}
}
\value{
data.frame
}
\description{
Generate correlation ellipse data
}
\examples{
cor_ellipse(r = .75)
}
