% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cor_heat}
\alias{cor_heat}
\title{Title}
\usage{
cor_heat(d, test_names = colnames(d), margins = 0.55, text_size = 4,
  ...)
}
\arguments{
\item{d}{Data}

\item{test_names}{Vector of names of variables}

\item{margins}{Width of margins for labels}

\item{text_size}{Size of text}

\item{...}{Additional parameters passed to superheat::superheat}
}
\description{
Title
}
