% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cor_text}
\alias{cor_text}
\title{Format text of a correlation matrix}
\usage{
cor_text(r, digits = 2, remove_leading_zero = TRUE, remove_diagonal = FALSE)
}
\arguments{
\item{r}{a matrix of correlations}

\item{digits}{Number of digits to round}

\item{remove_leading_zero}{If TRUE, remove leading zero from all correlations}

\item{remove_diagonal}{If TRUE, remove diagonal of ones}
}
\value{
a matrix of correlations as text
}
\description{
Format text of a correlation matrix
}
\examples{
cor_text(matrix(.5,2,2))
}
