% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cornames<-}
\alias{cornames<-}
\title{Name for square arrays like correlation matrices}
\usage{
cornames(x) <- value
}
\arguments{
\item{x}{a square array, matrix, or data.frame}

\item{value}{a vector of names}
}
\value{
a named square array, matrix, or data.frame
}
\description{
Name for square arrays like correlation matrices
}
\examples{
R <- tri2cor(.5)
cornames(R) <- c("A", "B")
}
