% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cross_vectors}
\alias{cross_vectors}
\title{Create unique combinations of vectors}
\usage{
cross_vectors(..., sep = "_")
}
\arguments{
\item{...}{vectors}

\item{sep}{Separate character}
}
\value{
A character vector
}
\description{
Create unique combinations of vectors
}
\examples{
cross_vectors(c("a", "b"), c(1,2))
}
