% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{df2matrix}
\alias{df2matrix}
\title{Convert data.frame and tibbles to matices with named rows and columns}
\usage{
df2matrix(d, first_col_is_row_names = TRUE)
}
\arguments{
\item{d}{data.frame or tibble}

\item{first_col_is_row_names}{TRUE if first column has row names}
}
\value{
matrix
}
\description{
Convert data.frame and tibbles to matices with named rows and columns
}
\examples{

d <- data.frame(rname = c("x1", "x2"), x1 = c(1,2), x2 = c(3,4))
df2matrix(d)
}
