% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{geom_richlabel}
\alias{geom_richlabel}
\title{A wrapper for ggtext::geom_richtext}
\usage{
geom_richlabel(
  mapping,
  label.margin = unit(2, "mm"),
  label.padding = unit(0.5, "mm"),
  label.color = NA,
  fill = "white",
  text_size = 12,
  ...
)
}
\arguments{
\item{mapping}{aes()}

\item{label.margin}{grid unit margin}

\item{label.padding}{grid unit margin}

\item{label.color}{color}

\item{fill}{color}

\item{text_size}{text size in point units}

\item{...}{additional parameters passed to ggtext::geom_richtext}
}
\value{
ggtext::geom_richtext
}
\description{
A wrapper for ggtext::geom_richtext
}
\examples{
library(tibble)
library(ggplot2)
library(dplyr)
library(WJSmisc)
tibble(x = 0, y = 0, l = "A") \%>\%
  ggplot() +
  geom_richlabel(aes(x,y,label = l))
}
