% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_quote}
\alias{get_quote}
\title{Retrieve text within a div with a named id}
\usage{
get_quote(id, file, blockquote = TRUE)
}
\arguments{
\item{id}{the id string (without #).}

\item{file}{file name}

\item{blockquote}{return as block quote (defaults to TRUE)}
}
\value{
a character vector of length 1
}
\description{
Retrieve text within a div with a named id
}
