% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ggsvg}
\alias{ggsvg}
\title{Save ggplot as .pdf, then convert to .svg via pdf2svg}
\usage{
ggsvg(f, width = 6.5, height = 6.5, ...)
}
\arguments{
\item{f}{Filename of the svg file. Omit the ".svg" at the end.}

\item{width}{width passed to ggplot2::ggsave}

\item{height}{width passed to ggplot2::ggsave}

\item{...}{Additional parameters passed to ggplot2::ggsave}
}
\value{

}
\description{
Save ggplot as .pdf, then convert to .svg via pdf2svg
}
