% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ggtext_size}
\alias{ggtext_size}
\title{Convert ggplot theme font size to geom_text size}
\usage{
ggtext_size(base_size, ratio = 0.8)
}
\arguments{
\item{base_size}{theme font size}

\item{ratio}{ratio of text size to theme font size. Defaults to .8 so that geom text will be the same size as default sized axis labels.}
}
\value{
numeric vector
}
\description{
Convert ggplot theme font size to geom_text size
}
\examples{

ggtext_size(16)
}
