% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{insert_latex_color}
\alias{insert_latex_color}
\title{Insert latex colors into a matrix}
\usage{
insert_latex_color(
  m,
  color_cells,
  color_rownames = NULL,
  color_colnames = NULL
)
}
\arguments{
\item{m}{matrix of values}

\item{color_cells}{matrix of latex colors for matrix cells}

\item{color_rownames}{vector of latex colors for row names}

\item{color_colnames}{vector of latex colors for column names}
}
\value{
matrix
}
\description{
Insert latex colors into a matrix
}
\examples{
# A matrix of zeros
m <- matrix(0, nrow = 2, ncol = 2)

# A matrix of NA values the same size as m
latex_colors <- m * NA

# Make the top row red
latex_colors[1,] <- "red"

# Insert colors into m
insert_latex_color(m, latex_colors)
}
