% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{latexarray}
\alias{latexarray}
\title{Make latex array}
\usage{
latexarray(
  M,
  left = "",
  right = "",
  env = "array",
  includenames = TRUE,
  align = NULL,
  lines = TRUE
)
}
\arguments{
\item{M}{A matrix}

\item{left}{left delimiter options: [(|}

\item{right}{right delimiter options: ])|}

\item{env}{Array environment}

\item{includenames}{Include column and row names}

\item{align}{Column alignment. For a three column matrix, alignment defaults to centering columns("ccc"). If there are row labels, the default would be "r|ccc" to right-align the row labels and separate them with vertical line.}

\item{lines}{Include lines separating column and row names}
}
\value{

}
\description{
Make latex array
}
\examples{
M <- diag(3)
colnames(M) <- LETTERS[1:3]
rownames(M) <- LETTERS[1:3]
latexarray(M)
}
