% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{lm_matrix}
\alias{lm_matrix}
\title{Regression from correlation matrix}
\usage{
lm_matrix(R, ind, dep)
}
\arguments{
\item{R}{correlation matrix}

\item{ind}{independent variables}

\item{dep}{dependent variable}
}
\value{
list of coefficients and R2
}
\description{
Regression from correlation matrix
}
