% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{make_clusters}
\alias{make_clusters}
\title{Create clustered data structures}
\usage{
make_clusters(
  data,
  .cluster_id = "cluster_id",
  .cluster_size = "cluster_size",
  .id = "id"
)
}
\arguments{
\item{data}{a vector of level-2 sample sizes or a data.frame with level-2 data which includes dsample sizes in one of its columns}

\item{.cluster_id}{name of cluster id variable in level-2 data}

\item{.cluster_size}{name of cluster size variable in level-2 data}

\item{.id}{name of new level-1 id variable}
}
\value{

}
\description{
Create clustered data structures
}
\examples{
# Number of clusters
k <- 20
# Level-2 data with cluster id, cluster size, and a level-2 covariate
d_l2 <- data.frame(
  classroom_id = 1:k,
  classroom_size = rpois(k, lambda = 30),
  Xj = rnorm(k))

# Make level-1 data frame
d_l1 <- make_clusters(
  data = d_l2,
  .cluster_id = "classroom_id",
  .cluster_size = "classroom_size",
  .id = "student_id")
d_l1
}
