% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{make_indicators}
\alias{make_indicators}
\title{Make latent variable indicators with rbeta_ms function}
\usage{
make_indicators(
  latent,
  indicators = NULL,
  mu = 0.8,
  sigma = 0.05,
  k = 3,
  digits = 3
)
}
\arguments{
\item{latent}{name of latent variable}

\item{indicators}{vector of indicator names (assigned automatically if left NULL)}

\item{mu}{mean of standardized coefficients}

\item{sigma}{sd of standardized coeficients}

\item{k}{number of indicator variables}

\item{digits}{number of digits to round coefficients}
}
\value{
lavaan code for latent variable assignment
}
\description{
Make latent variable indicators with rbeta_ms function
}
\examples{
make_indicators("depression", mu = 0.8, sigma = 0.05, k = 4)
}
