% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{modregplot}
\alias{modregplot}
\title{modregplot}
\usage{
modregplot(
  predictor_range = c(-4, 4),
  moderator_values = c(-1, 0, 1),
  intercept = 0,
  predictor_coef = 0,
  moderator_coef = 0,
  interaction_coef = 0,
  predictor_label = "X",
  criterion_label = "Y",
  moderator_label = "Moderator"
)
}
\arguments{
\item{predictor_range}{a length 2 vector of the range of values to be plotted on the predictor variable, Default: c(-4, 4)}

\item{moderator_values}{a vector of moderator values to be plotted, Default: c(-1, 0, 1)}

\item{intercept}{the intercept of the regression equation, Default: 0}

\item{predictor_coef}{the regression coefficient for the predictor variable, Default: 0}

\item{moderator_coef}{the regression coefficient for the moderator variable, Default: 0}

\item{interaction_coef}{the interaction term coefficent, Default: 0}

\item{predictor_label}{the label for the predictor variable, Default: 'X'}

\item{criterion_label}{the label for the moderator variable, Default: 'Y'}

\item{moderator_label}{PARAM_DESCRIPTION, Default: 'Moderator'}
}
\value{
a ggplot of the simple slopes
}
\description{
generates simple slopes plot from moderated regression equation
}
\examples{
modregplot(
  predictor_range = c(-2, 2),
  moderator_values = c(Low = -1, High = 1),
  intercept = 6,
  predictor_coef = 2,
  moderator_coef = 0,
  interaction_coef = 1,
  predictor_label = "Psychopathy",
  criterion_label = "Aggression",
  moderator_label = "Impulsivity"
)
}
