% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{multivariate_ci}
\alias{multivariate_ci}
\title{General a multivariate confidence interval for a set of scores}
\usage{
multivariate_ci(x, rxx, mu, sigma, ci = 0.95, v_names = names(x))
}
\arguments{
\item{x}{a vector of scores}

\item{rxx}{a vector reliability coefficients}

\item{mu}{a vector means}

\item{sigma}{a covariance matrix}

\item{ci}{confidence level}

\item{v_names}{a vector of names}
}
\value{
data.frame
}
\description{
General a multivariate confidence interval for a set of scores
}
\examples{
x_wisc <- c(
  vci = 130,
  vsi = 130,
  fri = 70,
  wmi = 130,
  psi = 130
)
rxx_wisc <- c(
  vci = .92,
  vsi = .92,
  fri = .93,
  wmi = .92,
  psi = .88
  )
R_wisc <- ("
  index	vci 	vsi 	fri 	wmi 	psi
  vci  	1.00	0.59	0.59	0.53	0.30
  vsi  	0.59	1.00	0.62	0.50	0.36
  fri  	0.59	0.62	1.00	0.53	0.31
  wmi  	0.53	0.50	0.53	1.00	0.36
  psi  	0.30	0.36	0.31	0.36	1.00") |>
    readr::read_tsv() |>
    tibble::column_to_rownames("index") |>
    as.matrix()
 multivariate_ci(
   x = x_wisc,
   rxx = rxx_wisc,
   mu = rep(100, 5),
   sigma = R_wisc * 225
 )
}
