% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{parallel_analysis}
\alias{parallel_analysis}
\title{ggplot of parallel analysis from the psych package}
\usage{
parallel_analysis(
  d,
  fm = "pa",
  vcolors = c("firebrick", "royalblue"),
  font_family = "serif",
  ...
)
}
\arguments{
\item{d}{data to be analyzed}

\item{fm}{factor method passed to psych::fa.parallel}

\item{vcolors}{vector of 2 colors for lines}

\item{font_family}{Name of font}

\item{...}{parameters passed to psych::fa.parallel}
}
\description{
ggplot of parallel analysis from the psych package
}
\examples{
d <- psych::bfi[,1:25]
parallel_analysis(d)
}
