% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{paste_matrix_from_clipboard}
\alias{paste_matrix_from_clipboard}
\title{Paste matrix code from clipboard}
\usage{
paste_matrix_from_clipboard(digits = 2, as_matrix = TRUE)
}
\arguments{
\item{digits}{Number of digits to round}

\item{as_matrix}{Convert to matrix. Defaults to `TRUE`}
}
\description{
Paste matrix code from clipboard
}
