% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{pdf2svg}
\alias{pdf2svg}
\title{Function that converts pdf files to svg. Must have the pdf2svg program installed (https://github.com/dawbarton/pdf2svg)}
\usage{
pdf2svg(f)
}
\arguments{
\item{f}{Filename of the pdf file. Omit the ".pdf" at the end.}
}
\description{
Function that converts pdf files to svg. Must have the pdf2svg program installed (https://github.com/dawbarton/pdf2svg)
}
