% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot_loading}
\alias{plot_loading}
\title{Creats a plot of factor loadings}
\usage{
plot_loading(f, factor_names = sort(colnames(f$loadings)))
}
\arguments{
\item{f}{output of a factor analysis from the psych::fa function}

\item{factor_names}{names of the factors #' @import patchwork}
}
\description{
Creats a plot of factor loadings
}
\examples{
library(psych)
fit <- fa(psych::bfi[,1:25], nfactors = 5)
plot_loading(fit)
}
