% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{prob_label}
\alias{prob_label}
\title{Format numeric probabilities as text labels}
\usage{
prob_label(
  p,
  accuracy = 0.01,
  digits = NULL,
  remove_leading_zero = TRUE,
  round_zero_one = TRUE
)
}
\arguments{
\item{p}{numeric vector of probabilities}

\item{accuracy}{accuracy of rounding}

\item{digits}{optional number of digits to round. Overrides accuracy parameter}

\item{remove_leading_zero}{Removes leading zero from probability}

\item{round_zero_one}{Apply rounding to 0 and 1}
}
\value{
character vector
}
\description{
Format numeric probabilities as text labels
}
\examples{
prob_label(seq(0,1, 0.1))
}
