% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{proportion2percentile}
\alias{proportion2percentile}
\title{Rounds proportions to significant digits both near 0 and 1, then converts to percentiles}
\usage{
proportion2percentile(
  p,
  digits = 2,
  remove_leading_zero = TRUE,
  add_percent_character = FALSE
)
}
\arguments{
\item{p}{probabiity}

\item{digits}{rounding digits}

\item{remove_leading_zero}{If TRUE, remove leading zero}

\item{add_percent_character}{If TRUE, add percent character}
}
\value{
chracter vector
}
\description{
Rounds proportions to significant digits both near 0 and 1, then converts to percentiles
}
\examples{
proportion2percentile(0.011111)
}
