% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{proportion_round}
\alias{proportion_round}
\title{Rounds proportions to significant digits both near 0 and 1}
\usage{
proportion_round(p, digits = 2)
}
\arguments{
\item{p}{probabiity}

\item{digits}{rounding digits}
}
\value{
numeric vector
}
\description{
Rounds proportions to significant digits both near 0 and 1
}
\examples{
proportion_round(0.01111)
}
