% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{pthreshold}
\alias{pthreshold}
\title{Probability a true score will be below a threshold}
\usage{
pthreshold(x, threshold, rxx, mu = 0, sigma = 1)
}
\arguments{
\item{x}{observed score}

\item{threshold}{threshold}

\item{rxx}{reliability coefficient}

\item{mu}{population mean}

\item{sigma}{population standard deviation}
}
\value{
probability
}
\description{
Probability a true score will be below a threshold
}
\examples{
pthreshold(x = .5, threshold = 1, rxx = 0.9)
}
