% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{pvalueAPA}
\alias{pvalueAPA}
\title{APA p-value rounding}
\usage{
pvalueAPA(p, inline = FALSE, mindigits = 2, maxdigits = 3)
}
\arguments{
\item{p}{probabiity}

\item{inline}{to be used in an inline rmarkdown (default is FALSE)}

\item{mindigits}{minimum rounding digits}

\item{maxdigits}{maximum rounding digits}
}
\value{
character
}
\description{
APA p-value rounding
}
\examples{
pvalueAPA(0.01111)
}
