% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{rbeta_ms}
\alias{rbeta_ms}
\title{Random beta distribution with specified mean and sd}
\usage{
rbeta_ms(n = 1, mu = 0.5, sigma = 0.025)
}
\arguments{
\item{n}{Number of data points}

\item{mu}{Mean of random beta distribution}

\item{sigma}{SD of random beta distribution}
}
\value{
a vector of numeric values
}
\description{
Random beta distribution with specified mean and sd
}
\examples{
rbeta_ms(n = 5, mu = 0.8, sigma = 0.1)
}
