% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{remove_leading_zero}
\alias{remove_leading_zero}
\title{Remove leading zero from numbers}
\usage{
remove_leading_zero(x)
}
\arguments{
\item{x}{vector of  numbers}
}
\value{
vector of characters
}
\description{
Remove leading zero from numbers
}
\examples{
remove_leading_zero(c(0.5,-0.2))
}
