% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{rotate2dmatrix}
\alias{rotate2dmatrix}
\title{Rotate a 2-column matrix}
\usage{
rotate2dmatrix(x, theta, degrees = FALSE, origin = c(0, 0))
}
\arguments{
\item{x}{a 2-column matrix}

\item{theta}{angle}

\item{degrees}{if TRUE, theta is in degrees instead of radians}

\item{origin}{point of rotation}
}
\value{
a rotated 2-column matrix
}
\description{
Rotate a 2-column matrix
}
\examples{
x <- matrix(seq(10), ncol = 2)
rotate2dmatrix(x, pi)
}
