% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{skewed_axis}
\alias{skewed_axis}
\title{Draw skewed axes in ggplot2}
\usage{
skewed_axis(
  theta,
  axis_title = "X",
  draw_ticks = TRUE,
  draw_axis_text = TRUE,
  remove_origin = TRUE,
  tick_height = 0.02,
  lwd = 0.5,
  text_size = 12,
  color = NA,
  family = NA,
  mu = 0,
  sigma = 1,
  tick_label_interval = ifelse(sigma\%\%3 == 0, 3, 2)
)
}
\arguments{
\item{theta}{angle in radians}

\item{axis_title}{character}

\item{draw_ticks}{logical}

\item{draw_axis_text}{logical}

\item{remove_origin}{logical}

\item{tick_height}{height of ticks}

\item{lwd}{line width}

\item{text_size}{size of text}

\item{color}{color of lines}

\item{family}{font family}

\item{mu}{mean of variable}

\item{sigma}{standard deviation of variable}

\item{tick_label_interval}{interval of ticks}
}
\value{
plot
}
\description{
Draw skewed axes in ggplot2
}
\examples{
library(ggplot2)
ggplot(data.frame(x = c(0,1), y = c(4,4))) +
  skewed_axis(pi / 2, color = "black") +
  skewed_axis(pi / 4, color = "black") +
  skewed_axis(0, color = "black") +
  theme_void() +
  coord_equal()
}
