% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{snake2subscript}
\alias{snake2subscript}
\title{Convert snake case to subscript}
\usage{
snake2subscript(
  x,
  sep = "_",
  prefix = "<sub>",
  suffix = "</sub>",
  collapse = NULL,
  recycle0 = FALSE
)
}
\arguments{
\item{x}{character}

\item{sep}{separator defaults to "_"}

\item{prefix}{character prefix}

\item{suffix}{character suffix}

\item{collapse}{character parameter passed to paste0}

\item{recycle0}{parameter passed to paste0}
}
\value{
a character
}
\description{
Convert snake case to subscript
}
\examples{
snake2subscript("x_1")
}
