% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{tri2cor}
\alias{tri2cor}
\title{Create square correlation matrix from lower triangle}
\usage{
tri2cor(x)
}
\arguments{
\item{x}{vector of correlations}
}
\value{
square matrix
}
\description{
Create square correlation matrix from lower triangle
}
\examples{
tri2cor(c(.2,.3,.4))
}
