% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{vector_angle}
\alias{vector_angle}
\title{Compute the angle of a vector}
\usage{
vector_angle(x, origin = c(0, 0), degrees = FALSE, allow_negative = FALSE)
}
\arguments{
\item{x}{A length-2 vector}

\item{origin}{A length-2 vector}

\item{degrees}{If TRUE, returns angles in degrees instead of radians}

\item{allow_negative}{If TRUE, returns angles between -pi and pi (-180 and +180. If FALSE, returns angles between 0 and 2 * pi (0 and 360)}
}
\value{
A length-1 vector
}
\description{
Compute the angle of a vector
}
\examples{
vector_angle(c(1,1))
vector_angle(c(1,1), degrees = TRUE)
}
