% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{composite_covariance}
\alias{composite_covariance}
\title{Computes covariances of composite scores given a covariance matrix and a weight matrix}
\usage{
composite_covariance(Sigma, w, correlation = FALSE)
}
\arguments{
\item{Sigma}{Covariance matrix}

\item{w}{Weight matrix. Must have the same number of rows as R}

\item{correlation}{If TRUE, return correlations instead of covariances}
}
\description{
Computes covariances of composite scores given a covariance matrix and a weight matrix
}
\examples{
# Create covariance matrix
Sigma <- matrix(0.6, nrow = 5, ncol = 5)
diag(Sigma) <- 1
# Create weight matrix
w <- matrix(0, nrow = 5, ncol = 2)
w[1:2,1] <- 1
w[3:5,2] <- 1
w
# covariance matrix of weighted sums
composite_covariance(Sigma, w)
}
