% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot_loading}
\alias{plot_loading}
\title{Creats a plot of factor loadings}
\usage{
plot_loading(
  f,
  font_family = "serif",
  font_size = 14,
  factor_names = sort(colnames(f$loadings)),
  nudge_loadings = 0.05
)
}
\arguments{
\item{f}{output of a factor analysis from the psych::fa function}

\item{font_family}{Name of font}

\item{font_size}{Size of font}

\item{factor_names}{names of the factors #'}

\item{nudge_loadings}{nudge loadings on x dimension}
}
\description{
Creats a plot of factor loadings
}
\examples{
library(psych)
fit <- fa(psych::bfi[,1:25], nfactors = 5)
plot_loading(fit)
}
